+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Corrected outdated `SearchQueryBloc` documentation
 
+### Changed
+- Updated `blocTest` and `whenListen` typing
+
 ## [0.3.0] - 2020-08-08
 ### Changed
 - **Breaking:** `SearchQueryBloc` is now `SearchQueryCubit`